refactor(jsonpath): extract the strict-subset projector so two callers share it
Pure move, no behaviour change. Separated from the feature that needs it so the
diff that follows is only the feature.

The strict-subset JSONPath parser lived unexported inside internal/api/webhook,
which pulls loop, store, lookup, runner and otel. The team-workflow variable
binding needs the SAME grammar from internal/teamrun, a package that keeps a
deliberately small dependency set — so the parser moves to a stdlib-only
internal/jsonpath and webhook calls it.

Copying it instead would have been smaller and worse. This grammar is a security
boundary, not a convenience: the paths come from operator definitions projecting
ATTACKER-INFLUENCEABLE documents (an inbound webhook body; soon a channel
message), and its whole job is making wildcards, recursive descent, filters and
script expressions UNREACHABLE. Two copies of that are two things to keep
rejecting the same set, and the day they disagree is the day one caller starts
evaluating what the other refuses.

The two grammar tests move with the grammar, including the reject list — that
list is the point of the package, so it belongs beside it rather than beside one
of its callers.

TESTED: `go test ./...` clean; internal/jsonpath and internal/api/webhook both
pass with the tests unchanged apart from the renamed call.
