+++ internal/api/webhook/payload.go
@@ -3,8 +3,8 @@ package webhook
 import (
 	"encoding/json"
 	"fmt"
-	"strconv"
-	"strings"
+
+	"github.com/denn-gubsky/loomcycle/internal/jsonpath"
 )
 
 // projectResult is the outcome of applying a WebhookDef's payload_mapping
@@ -58,148 +58,17 @@ func projectPayload(mapping map[string]string, body []byte) (projectResult, erro
 	}
 
 	for target, path := range mapping {
-		segs, err := parsePath(path)
+		segs, err := jsonpath.Parse(path)
 		if err != nil {
 			return projectResult{}, fmt.Errorf("payload_mapping[%q]: %w", target, err)
 		}
-		v, ok := evalPath(doc, segs)
+		v, ok := jsonpath.Eval(doc, segs)
 		if !ok {
 			res.Fields[target] = ""
 			res.MissingKeys = append(res.MissingKeys, target)
 			continue
 		}
-		res.Fields[target] = stringify(v)
+		res.Fields[target] = jsonpath.Stringify(v)
 	}
 	return res, nil
-}
-
-// pathSeg is one resolved step in a parsed JSONPath: either a map key
-// (Key set, IsIndex false) or an array index (Index set, IsIndex true).
-type pathSeg struct {
-	Key     string
-	Index   int
-	IsIndex bool
-}
-
-// parsePath validates and tokenizes a strict-subset JSONPath. Returns an
-// error for any shape outside the allowlist (wildcards, filters, recursive
-// descent, empty segments). The grammar is intentionally tiny so the
-// rejection surface is exhaustive: a leading "$", then zero or more
-// segments of the form `.key` or `[N]`.
-func parsePath(path string) ([]pathSeg, error) {
-	path = strings.TrimSpace(path)
-	if path == "" {
-		return nil, fmt.Errorf("empty path")
-	}
-	if path[0] != '$' {
-		return nil, fmt.Errorf("path must start with $")
-	}
-	// Reject recursive descent ("..") and the bare-wildcard forms outright,
-	// before the segment walk, so the error message names the actual
-	// violation rather than a downstream parse glitch.
-	if strings.Contains(path, "..") {
-		return nil, fmt.Errorf("recursive descent (..) not supported")
-	}
-	if strings.Contains(path, "*") {
-		return nil, fmt.Errorf("wildcard (*) not supported")
-	}
-	if strings.Contains(path, "?") || strings.Contains(path, "@") {
-		return nil, fmt.Errorf("filter expressions not supported")
-	}
-
-	rest := path[1:] // strip the leading $
-	var segs []pathSeg
-	for len(rest) > 0 {
-		switch rest[0] {
-		case '.':
-			rest = rest[1:]
-			// Read a key up to the next '.' or '['.
-			end := strings.IndexAny(rest, ".[")
-			var key string
-			if end == -1 {
-				key = rest
-				rest = ""
-			} else {
-				key = rest[:end]
-				rest = rest[end:]
-			}
-			if key == "" {
-				return nil, fmt.Errorf("empty key segment")
-			}
-			segs = append(segs, pathSeg{Key: key})
-		case '[':
-			end := strings.IndexByte(rest, ']')
-			if end == -1 {
-				return nil, fmt.Errorf("unterminated [ index")
-			}
-			idxStr := rest[1:end]
-			idx, err := strconv.Atoi(strings.TrimSpace(idxStr))
-			if err != nil || idx < 0 {
-				// Only non-negative integer indices are allowed. A quoted
-				// key form (['key']) is intentionally NOT supported — it
-				// widens the grammar with no payload-mapping need.
-				return nil, fmt.Errorf("invalid array index %q", idxStr)
-			}
-			segs = append(segs, pathSeg{Index: idx, IsIndex: true})
-			rest = rest[end+1:]
-		default:
-			return nil, fmt.Errorf("unexpected character %q in path", string(rest[0]))
-		}
-	}
-	return segs, nil
-}
-
-// evalPath walks the parsed segments over a decoded JSON document. Returns
-// (value, true) when every segment resolves; (nil, false) on any miss
-// (wrong type, absent key, out-of-range index). No panics — a type
-// mismatch is a miss, not a crash.
-func evalPath(doc interface{}, segs []pathSeg) (interface{}, bool) {
-	cur := doc
-	for _, s := range segs {
-		if s.IsIndex {
-			arr, ok := cur.([]interface{})
-			if !ok || s.Index >= len(arr) {
-				return nil, false
-			}
-			cur = arr[s.Index]
-			continue
-		}
-		obj, ok := cur.(map[string]interface{})
-		if !ok {
-			return nil, false
-		}
-		v, present := obj[s.Key]
-		if !present {
-			return nil, false
-		}
-		cur = v
-	}
-	return cur, true
-}
-
-// stringify converts a resolved JSON value to the flat string the mapping
-// produces. Strings pass through verbatim; numbers/bools render via their
-// natural form; objects/arrays/null render as compact JSON so a mapping
-// that targets a sub-object still yields a deterministic string rather than
-// Go's %v formatting.
-func stringify(v interface{}) string {
-	switch t := v.(type) {
-	case string:
-		return t
-	case float64:
-		// json.Unmarshal decodes all numbers to float64. strconv with -1
-		// precision avoids trailing zeros and scientific notation for the
-		// common integer-id case.
-		return strconv.FormatFloat(t, 'f', -1, 64)
-	case bool:
-		return strconv.FormatBool(t)
-	case nil:
-		return ""
-	default:
-		b, err := json.Marshal(t)
-		if err != nil {
-			return ""
-		}
-		return string(b)
-	}
 }
